Custom chalenges - Challenges are now controlled from a json file and can be manipulated. Will be used for steam workshop eventually - Stats screen has been updated - Avatars added for all leaderboards - Key sprite updated with correct position when collected - Fixed level editor liquids where when selected in the item menu they would be started in placement mode - Fixed level editor liquids where you could place them in play mode

+++ scripts/scr_savechallengetime/scr_savechallengetime.gml
@@ -1,28 +1,20 @@
 function scr_savechallengetime() {
 	if global.cheats = 0 {
-	var directory = directory_set("//Save Files/")
-	
-	if (file_exists(directory + "ChallengeTime.sav")) file_delete(directory + "ChallengeTime.sav");
-	ini_open(directory + "ChallengeTime.sav");
-	ini_write_real("Challenges","Tutorial Challenge",global.tutorialchallengetime);
-	ini_write_real("Challenges","Ladder Challenge",global.ladderchallengetime);
-	ini_write_real("Challenges","Big Room Challenge",global.bigroomchallengetime);
-	ini_write_real("Challenges","Slippery Challenge",global.slipperychallengetime);
-	ini_write_real("Challenges","Blind Challenge",global.blindchallengetime);
-	ini_write_real("Challenges","Troop Challenge",global.troopchallengetime);
-	ini_write_real("Challenges","Speed Challenge",global.speedchallengetime);
-	ini_write_real("Challenges","Spike Challenge",global.spikechallengetime);
-	ini_write_real("Challenges","Kaizo Challenge",global.kaizochallengetime);
-	ini_write_real("Challenges","World 6 Challenge",global.world6challengetime);
-	ini_write_real("Challenges","Water Challenge",global.waterchallengetime);
-	ini_write_real("Challenges","Moving Challenge",global.movingchallengetime);
-	ini_write_real("Challenges","Community Challenge",global.communitychallengetime);
-	ini_write_real("Challenges","D.Jump Challenge",global.djumpchallengetime);
-	ini_write_real("Challenges","C.Spike Challenge",global.cspikechallengetime);
-	ini_write_real("Challenges","World 7 Challenge",global.world7challengetime);
-	ini_write_real("Challenges","Invisible Challenge",global.invisiblechallengetime);
-	ini_write_real("Challenges","Breakable Challenge",global.breakablechallengetime);
-	
-	ini_close();
+		if (!variable_global_exists("challenge_defs")) return;
+		if (!ds_exists(global.challenge_times, ds_type_map)) global.challenge_times = ds_map_create();
+		
+		var directory = directory_set("//Save Files/")
+		
+		if (file_exists(directory + "ChallengeTime.sav")) file_delete(directory + "ChallengeTime.sav");
+		ini_open(directory + "ChallengeTime.sav");
+		var _keys = scr_map_keys(global.challenge_defs);
+		for (var i = 0; i < ds_list_size(_keys); i++) {
+			var _key = _keys[| i];
+			var _def = global.challenge_defs[? _key];
+			var _value = scr_challenge_get_time(_def.id);
+			ini_write_real("Challenges", _def.save_key, _value);
+		}
+		ds_list_destroy(_keys);
+		ini_close();
 	}
 }
